% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_distance.R
\name{compile_distance}
\alias{compile_distance}
\title{Compile the functions for the distance example.}
\usage{
compile_distance(module = Module())
}
\arguments{
\item{module}{an existing LLVM module to use for compilation, optional}
}
\value{
An LLVM module containing the compiled functions.
}
\description{
Compile the functions for the distance example.
}

