% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_knn.R
\name{compile_knn}
\alias{compile_knn}
\title{Compile the functions for the knn example.}
\usage{
compile_knn(module = Module())
}
\arguments{
\item{module}{an existing LLVM module to use for compilation, optional}
}
\value{
An LLVM module containing the compiled functions.
}
\description{
Compile the functions for the knn example.
}

