% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_distance.R
\name{distance}
\alias{.distance}
\alias{distance}
\title{Compute distances from columns of x to columns of y.}
\usage{
.distance(x, y, nrow, ncol_x, ncol_y, p)
}
\arguments{
\item{x}{numeric, first matrix}

\item{y}{numeric, second matrix}

\item{nrow}{integer, number of rows in x (or y)}

\item{ncol_x}{integer, number of columns in x}

\item{ncol_y}{integer, number of columns in y}

\item{p}{numeric, Minkowski parameter (e.g., 2 gives Euclidean distance)}
}
\description{
This version is non-idiomatic to suit the compiler's quirks, and should
not be called directly in the R interpreter.
}

