% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_knn.R
\name{knn}
\alias{.knn}
\alias{knn}
\title{Classify points using k-nearest neighbors.}
\usage{
.knn(distances, labels, n_class, k, n_train, n_test)
}
\arguments{
\item{distances}{numeric, distance matrix, where rows correspond to
training points and columns correspond to test points}

\item{labels}{integer, class labels starting at 1}

\item{n_class}{integer, number of classes}

\item{k}{integer, number of nearest neighbors to use}

\item{n_train}{integer, number of training points}

\item{n_test}{integer, number of test points}
}
\description{
This version is non-idiomatic to suit the compiler's quirks, and should
not be called directly in the R interpreter.
}

