% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_distance.R
\name{minkowski}
\alias{.minkowski}
\alias{minkowski}
\title{Compute the Minkowski distance between two vectors.}
\usage{
.minkowski(x, y, len, p)
}
\arguments{
\item{x}{numeric, first vector}

\item{y}{numeric, second vector}

\item{len}{integer, length of x and y}

\item{p}{numeric, Minkowski parameter (e.g., 2 gives Euclidean distance)}
}
\description{
Compute the Minkowski distance between two vectors.
}

