% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_knn.R
\name{order}
\alias{.order}
\alias{order}
\title{Compute an index into x which specifies the elements in ascending order.}
\usage{
.order(x, len)
}
\arguments{
\item{x}{numeric, vector of values}

\item{len}{integer, length of x}
}
\description{
This is a clone of R's `order()` function.
}

