% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCompilerExamples.R
\docType{package}
\name{RCompilerExamples}
\alias{RCompilerExamples}
\alias{RCompilerExamples-package}
\title{Examples for RLLVMCompile}
\description{
This package contains case studies demonstrating statistical applications of
Rllvm and RLLVMCompile.
}

