% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_knn.R
\name{which.max}
\alias{.which.max}
\alias{which.max}
\title{Compute index of max element.}
\usage{
.which.max(x, len)
}
\arguments{
\item{x}{integer, vector of values}

\item{len}{numeric, length of x}
}
\description{
This is a clone of R's `which.max()`.
}

